pyproject.toml
[project]
name = "lstm-frequency-extraction"
version = "1.0.0"
description = "Professional LSTM implementation for frequency extraction from mixed signals"
authors = [
    { name = "Professional ML Engineering Team", email = "ml-team@example.com" }
]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",
    "pyyaml>=6.0",
    "tensorboard>=2.13.0",
    "wandb>=0.15.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "hydra-core>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "isort>=5.12.0",
]

notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.23.0",
    "ipywidgets>=8.0.0",
]

[project.scripts]
lstm-freq-extract = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

